Break Stream into new file
