+++ Sources/ApolloWebSocket/WebSocket.swift
@@ -8,9 +8,7 @@
 //  Original Work License: http://www.apache.org/licenses/LICENSE-2.0
 //  Derived Work License: https://github.com/apollographql/apollo-ios/blob/main/LICENSE
 
-
 import Foundation
-import CoreFoundation
 import CommonCrypto
 
 //Standard WebSocket close codes
@@ -56,200 +54,9 @@ public struct SSLSettings {
   #endif
 }
 
-public protocol WSStreamDelegate: class {
-  func newBytesInStream()
-  func streamDidError(error: Error?)
-}
-
-//This protocol is to allow custom implemention of the underlining stream. This way custom socket libraries (e.g. linux) can be used
-public protocol WSStream {
-  var delegate: WSStreamDelegate? {get set}
-  func connect(url: URL, port: Int, timeout: TimeInterval, ssl: SSLSettings, completion: @escaping ((Error?) -> Void))
-  func write(data: Data) -> Int
-  func read() -> Data?
-  func cleanup()
-  #if os(Linux) || os(watchOS)
-  #else
-  func sslTrust() -> (trust: SecTrust?, domain: String?)
-  #endif
-}
-
-open class FoundationStream : NSObject, WSStream, StreamDelegate  {
-  private let workQueue = DispatchQueue(label: "com.vluxe.starscream.websocket", attributes: [])
-  private var inputStream: InputStream?
-  private var outputStream: OutputStream?
-  public weak var delegate: WSStreamDelegate?
-  let BUFFER_MAX = 4096
-
-  public var enableSOCKSProxy = false
-
-  public func connect(url: URL, port: Int, timeout: TimeInterval, ssl: SSLSettings, completion: @escaping ((Error?) -> Void)) {
-    var readStream: Unmanaged<CFReadStream>?
-    var writeStream: Unmanaged<CFWriteStream>?
-    let h = url.host! as NSString
-    CFStreamCreatePairWithSocketToHost(nil, h, UInt32(port), &readStream, &writeStream)
-    inputStream = readStream!.takeRetainedValue()
-    outputStream = writeStream!.takeRetainedValue()
-
-    #if os(watchOS) //watchOS us unfortunately is missing the kCFStream properties to make this work
-    #else
-    if enableSOCKSProxy {
-      let proxyDict = CFNetworkCopySystemProxySettings()
-      let socksConfig = CFDictionaryCreateMutableCopy(nil, 0, proxyDict!.takeRetainedValue())
-      let propertyKey = CFStreamPropertyKey(rawValue: kCFStreamPropertySOCKSProxy)
-      CFWriteStreamSetProperty(outputStream, propertyKey, socksConfig)
-      CFReadStreamSetProperty(inputStream, propertyKey, socksConfig)
-    }
-    #endif
-
-    guard let inStream = inputStream, let outStream = outputStream else { return }
-    inStream.delegate = self
-    outStream.delegate = self
-    if ssl.useSSL {
-      inStream.setProperty(StreamSocketSecurityLevel.negotiatedSSL as AnyObject, forKey: Stream.PropertyKey.socketSecurityLevelKey)
-      outStream.setProperty(StreamSocketSecurityLevel.negotiatedSSL as AnyObject, forKey: Stream.PropertyKey.socketSecurityLevelKey)
-      #if os(watchOS) //watchOS us unfortunately is missing the kCFStream properties to make this work
-      #else
-      var settings = [NSObject: NSObject]()
-      if ssl.disableCertValidation {
-        settings[kCFStreamSSLValidatesCertificateChain] = NSNumber(value: false)
-      }
-      if ssl.overrideTrustHostname {
-        if let hostname = ssl.desiredTrustHostname {
-          settings[kCFStreamSSLPeerName] = hostname as NSString
-        } else {
-          settings[kCFStreamSSLPeerName] = kCFNull
-        }
-      }
-      if let sslClientCertificate = ssl.sslClientCertificate {
-        settings[kCFStreamSSLCertificates] = sslClientCertificate.streamSSLCertificates
-      }
-
-      inStream.setProperty(settings, forKey: kCFStreamPropertySSLSettings as Stream.PropertyKey)
-      outStream.setProperty(settings, forKey: kCFStreamPropertySSLSettings as Stream.PropertyKey)
-      #endif
-
-      #if os(Linux)
-      #else
-      if let cipherSuites = ssl.cipherSuites {
-        #if os(watchOS) //watchOS us unfortunately is missing the kCFStream properties to make this work
-        #else
-        if let sslContextIn = CFReadStreamCopyProperty(inputStream, CFStreamPropertyKey(rawValue: kCFStreamPropertySSLContext)) as! SSLContext?,
-           let sslContextOut = CFWriteStreamCopyProperty(outputStream, CFStreamPropertyKey(rawValue: kCFStreamPropertySSLContext)) as! SSLContext? {
-          let resIn = SSLSetEnabledCiphers(sslContextIn, cipherSuites, cipherSuites.count)
-          let resOut = SSLSetEnabledCiphers(sslContextOut, cipherSuites, cipherSuites.count)
-          if resIn != errSecSuccess {
-            completion(WSError(type: .invalidSSLError, message: "Error setting ingoing cypher suites", code: Int(resIn)))
-          }
-          if resOut != errSecSuccess {
-            completion(WSError(type: .invalidSSLError, message: "Error setting outgoing cypher suites", code: Int(resOut)))
-          }
-        }
-        #endif
-      }
-      #endif
-    }
-
-    CFReadStreamSetDispatchQueue(inStream, workQueue)
-    CFWriteStreamSetDispatchQueue(outStream, workQueue)
-    inStream.open()
-    outStream.open()
-
-    var out = timeout// wait X seconds before giving up
-    workQueue.async { [weak self] in
-      while !outStream.hasSpaceAvailable {
-        usleep(100) // wait until the socket is ready
-        out -= 100
-        if out < 0 {
-          completion(WSError(type: .writeTimeoutError, message: "Timed out waiting for the socket to be ready for a write", code: 0))
-          return
-        } else if let error = outStream.streamError {
-          completion(error)
-          return // disconnectStream will be called.
-        } else if self == nil {
-          completion(WSError(type: .closeError, message: "socket object has been dereferenced", code: 0))
-          return
-        }
-      }
-      completion(nil) //success!
-    }
-  }
-
-  public func write(data: Data) -> Int {
-    guard let outStream = outputStream else {return -1}
-    let buffer = UnsafeRawPointer((data as NSData).bytes).assumingMemoryBound(to: UInt8.self)
-    return outStream.write(buffer, maxLength: data.count)
-  }
-
-  public func read() -> Data? {
-    guard let stream = inputStream else {return nil}
-    let buf = NSMutableData(capacity: BUFFER_MAX)
-    let buffer = UnsafeMutableRawPointer(mutating: buf!.bytes).assumingMemoryBound(to: UInt8.self)
-    let length = stream.read(buffer, maxLength: BUFFER_MAX)
-    if length < 1 {
-      return nil
-    }
-    return Data(bytes: buffer, count: length)
-  }
-
-  public func cleanup() {
-    if let stream = inputStream {
-      stream.delegate = nil
-      CFReadStreamSetDispatchQueue(stream, nil)
-      stream.close()
-    }
-    if let stream = outputStream {
-      stream.delegate = nil
-      CFWriteStreamSetDispatchQueue(stream, nil)
-      stream.close()
-    }
-    outputStream = nil
-    inputStream = nil
-  }
-
-  #if os(Linux) || os(watchOS)
-  #else
-  public func sslTrust() -> (trust: SecTrust?, domain: String?) {
-    guard let outputStream = outputStream else { return (nil, nil) }
-
-    let trust = outputStream.property(forKey: kCFStreamPropertySSLPeerTrust as Stream.PropertyKey) as! SecTrust?
-    var domain = outputStream.property(forKey: kCFStreamSSLPeerName as Stream.PropertyKey) as! String?
-    if domain == nil,
-       let sslContextOut = CFWriteStreamCopyProperty(outputStream, CFStreamPropertyKey(rawValue: kCFStreamPropertySSLContext)) as! SSLContext? {
-      var peerNameLen: Int = 0
-      SSLGetPeerDomainNameLength(sslContextOut, &peerNameLen)
-      var peerName = Data(count: peerNameLen)
-      let _ = peerName.withUnsafeMutableBytes { (peerNamePtr: UnsafeMutablePointer<Int8>) in
-        SSLGetPeerDomainName(sslContextOut, peerNamePtr, &peerNameLen)
-      }
-      if let peerDomain = String(bytes: peerName, encoding: .utf8), peerDomain.count > 0 {
-        domain = peerDomain
-      }
-    }
-
-    return (trust, domain)
-  }
-  #endif
-
-  /**
-   Delegate for the stream methods. Processes incoming bytes
-   */
-  open func stream(_ aStream: Stream, handle eventCode: Stream.Event) {
-    if eventCode == .hasBytesAvailable {
-      if aStream == inputStream {
-        delegate?.newBytesInStream()
-      }
-    } else if eventCode == .errorOccurred {
-      delegate?.streamDidError(error: aStream.streamError)
-    } else if eventCode == .endEncountered {
-      delegate?.streamDidError(error: nil)
-    }
-  }
-}
-
 //WebSocket implementation
 
-open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WSStreamDelegate {
+open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WebSocketStreamDelegate {
 
   public enum OpCode : UInt8 {
     case continueFrame = 0x0
@@ -261,11 +68,6 @@ open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WSStreamDelegat
     case pong = 0xA
     // B-F reserved.
   }
-
-  public static let ErrorDomain = "WebSocket"
-
-  // Where the callback is executed. It defaults to the main UI thread queue.
-  public var callbackQueue = DispatchQueue.main
 
   private struct Constants {
     static let headerWSUpgradeName     = "Upgrade"
@@ -315,6 +117,9 @@ open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WSStreamDelegat
   /// and also connection/disconnect messages.
   public weak var delegate: WebSocketClientDelegate?
 
+  // Where the callback is executed. It defaults to the main UI thread queue.
+  public var callbackQueue = DispatchQueue.main
+
   public var onConnect: (() -> Void)?
   public var onDisconnect: ((Error?) -> Void)?
   public var onText: ((String) -> Void)?
@@ -358,7 +163,7 @@ open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WSStreamDelegat
     var compressor:Compressor? = nil
   }
 
-  private var stream: WSStream
+  private var stream: WebSocketStream
   private var connected = false
   private var isConnecting = false
   private let mutex = NSLock()
@@ -379,9 +184,9 @@ open class WebSocket: NSObject, WebSocketClient, StreamDelegate, WSStreamDelegat
   }
 
   /// Used for setting protocols.
-  public init(request: URLRequest, protocols: [String]? = nil, stream: WSStream = FoundationStream()) {
+  public init(request: URLRequest, protocols: [String]? = nil) {
     self.request = request
-    self.stream = stream
+    self.stream = FoundationStream()
     if request.value(forHTTPHeaderField: Constants.headerOriginName) == nil {
       guard let url = request.url else {return}
       var origin = url.absoluteString
